+++ dsl/src/ir.rs
@@ -9,22 +9,7 @@ impl StepId {
   }
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub struct FutureId(pub String);
-
-impl FutureId {
-  pub fn new(id: impl Into<String>) -> Self {
-    Self(id.into())
-  }
-
-  // Construct a runtime FutureId from an IR FutureLabel and a fiber-specific unique id
-  pub fn from_label(
-    label: FutureLabel,
-    unique_id: u64,
-  ) -> Self {
-    Self(format!("{}_{}", label.0, unique_id))
-  }
-}
+// Note: Runtime `FutureId` moved into runtime modules.
 
 // IR-only identifier to label futures for awaits/links.
 // This is not used by the runtime which works with concrete `FutureId`s.
